requirements.txt
ccxt
python-telegram-bot
pandas
numpy
fastapi
uvicorn
python-dotenv
requests
jinja2